alx_travel_app/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
PyMySQL==1.1.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Pillow==10.1.0
django-filter==23.5
django-environ
requests
celery
